+++ src/react/components/button/XUIButton.js
@@ -4,7 +4,6 @@ import cn from 'classnames';
 import XUILoader from '../loader/XUILoader';
 import {
   buttonTypes,
-  iconSizeClassNames,
   sizeClassNames,
   variantClassNames,
   widthClassNames,
@@ -22,15 +21,6 @@ import SizeContext from '../../contexts/SizeContext';
  * @return {boolean} True if is a borderless button
  */
 const isBorderlessVariant = variant => variant.indexOf('borderless') > -1;
-
-/**
- * Returns true if the button is an icon variant
- *
- * @private
- * @param {string} variant - The button variant
- * @return {boolean} True if is an icon button
- */
-const isIconVariant = variant => variant.indexOf('icon') > -1;
 
 /**
  * Returns a class name for the button depending on the button variant string given. Will return
@@ -199,24 +189,22 @@ export default class XUIButton extends React.PureComponent {
             buttonChildren = loader;
           }
 
-          const isIconDependentClassNames = isIconVariant(variantClass)
-            ? cn(iconSizeClassNames[size])
-            : cn(
-                variant !== 'unstyled' && sizeClassNames[size],
-                widthClassNames[fullWidth],
-                isInverted &&
-                  (isBorderlessVariant(variantClass)
-                    ? `${ns}-button-borderless-inverted`
-                    : `${ns}-button-inverted`),
-                isGrouped && `${ns}-button-grouped`,
-                minLoaderWidth && `${ns}-button-min-loader-width`,
-              );
+          const combinedPropClassNames = cn(
+            variant !== 'unstyled' && sizeClassNames[size],
+            widthClassNames[fullWidth],
+            isInverted &&
+              (isBorderlessVariant(variantClass)
+                ? `${ns}-button-borderless-inverted`
+                : `${ns}-button-inverted`),
+            isGrouped && `${ns}-button-grouped`,
+            minLoaderWidth && `${ns}-button-min-loader-width`,
+          );
 
           const buttonClassNames = cn(
             `${ns}-button`,
             className,
             variantClass,
-            isIconDependentClassNames,
+            combinedPropClassNames,
             isDisabled && `${ns}-button-is-disabled`,
           );
 
@@ -302,7 +290,8 @@ XUIButton.propTypes = {
 
   /** Determines the styling variation to apply: `standard`, `primary`, `create`, `negative`, `link`,
    * 'borderless-standard', 'borderless-primary', 'borderless-create', 'borderless-negative',
-   * 'borderless-negative', 'icon', 'icon-inverted', or `unstyled`. */
+   * 'borderless-negative' or `unstyled`.
+   */
   variant: PropTypes.oneOf(Object.keys(variantClassNames)),
 
   /**